requirements.txt
# These are the explicit requirements we know about.
#
# We actually *do care* about their dependencies, because we must
# have the same enviroment in the virtualenv on the production server
# and in our development virtualenvs (unless you break your local one
# intentionaly...). Therefore the all implicit dependecies are listed
# bellow, as listed by `pip freeze -l`, and that should make sure we're
# using the same set of packages in the production and in the development
# as well.

# required by guest-setup module
ansible==2.2.2.0
# required to build docker images
ansible-container==0.9.2
# required by many modules for XML/HTML manipulatiion
beautifulsoup4==4.5.3
# optional - enables logging colors
colorama==0.3.7
# required by restraint modules
enum34==1.1.6
# required by jenkins module
jenkinsapi==0.3.3
# required by jenkins-jobs
jenkins-job-builder==1.6.2
# required by urlgrabber/covscan
pycurl==7.43.0
# required by opentack module - novaclient
python-novaclient==7.1.0
python-openstackclient==3.9.0
# required by koji module
pyOpenSSL==17.0.0
# required to load configuration
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.34
# required by sentry
raven==6.0.0
# required by many packages - we must list it explicitly to force version
# compatible with all their requirements
requests==2.12.5
# simplify usage of kerberos auth in requests
requests-kerberos==0.11.0
# required by openstack module
retrying==1.3.3
# required for documentation and help processing
Sphinx==1.5.2
sphinx-argparse==0.2.0
sphinxcontrib-programoutput==0.11
sphinx-rtd-theme==0.1.9
# required by notify-bus module
stomp.py==4.1.17
# required for better formatting
tabulate==0.7.7
# required by covscan module
urlgrabber==3.10.2
# required to normalize URLs
urlnorm==1.1.4


# Implicit requirementes
asn1crypto==0.22.0
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
cffi==1.10.0
cliff==2.4.0
cmd2==0.7.0
CommonMark==0.7.3
contextlib2==0.5.4
cryptography==1.8.1
debtcollector==1.12.0
deprecation==1.0
docker==2.5.1
docker-pycreds==0.2.1
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
idna==2.5
ipaddress==1.0.18
iso8601==0.1.11
Jinja2==2.9.6
jsonpatch==1.15
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.19.0
lxml==3.7.3
MarkupSafe==1.0
monotonic==1.3
msgpack-python==0.4.8
multi-key-dict==2.0.3
netaddr==0.7.19
netifaces==0.10.5
#nitrate==1.3.1
git+https://github.com/psss/python-nitrate.git@1.3-2#egg=nitrate
openstacksdk==0.9.14
os-client-config==1.26.0
osc-lib==1.3.0
oslo.config==3.23.0
oslo.i18n==3.14.0
oslo.serialization==2.17.0
oslo.utils==3.23.0
paramiko==2.1.2
pbr==2.0.0
positional==1.1.1
prettytable==0.7.2
proton==0.8.8
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pykerberos==1.1.14
python-cinderclient==2.0.1
python-dateutil==2.6.0
python-glanceclient==2.6.0
python-jenkins==0.4.14
python-keystoneclient==3.10.0
python-krbV==1.0.90
python-qpid-proton==0.17.0
pytz==2016.10
requestsexceptions==1.2.0
rfc3986==0.4.1
simplejson==3.10.0
stevedore==1.21.0
unicodecsv==0.14.1
warlock==1.2.0
websocket-client==0.40.0
wrapt==1.10.10
